requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-apscheduler==0.6.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
psycopg==3.1.17
psycopg-binary==3.1.17
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
twilio==8.12.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.4
